+++ app/src/main/java/com/konkuk/medicarecall/data/api/elders/HomeService.kt
@@ -1,7 +1,7 @@
-package com.konkuk.medicarecall.data.api
+package com.konkuk.medicarecall.data.api.elders
 
-import com.konkuk.medicarecall.data.dto.response.HomeResponseDto
 import com.konkuk.medicarecall.data.dto.request.ImmediateCallRequestDto
+import com.konkuk.medicarecall.data.dto.response.HomeResponseDto
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
